Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    float a,b,c,avg;
-    scanf("%f %f %f", &a,&b,&c);
+    int a,b,c,avg;
+    scanf("%d %d %d", &a,&b,&c);
     avg=a+b+c/3;
-    prinf("Average: %.2f", avg);
+    prinf("Average: %d", avg);
     return 0;
 }
